+++ test/utils_tests.c
@@ -17,7 +17,7 @@ static const char hex2[] = "AA969cdfFFffFF3bad960b0b000aca2ac329deea5c2328ebc6f2
 
 void test_utils()
     {
-    int outlen = 0;
+    size_t outlen = 0;
     unsigned char data[] = { 0x00, 0xFF, 0x00, 0xAA, 0x00, 0xFF, 0x00, 0xAA };
     char hash[] = "28969cdfa74a12c82f3bad960b0b000aca2ac329deea5c2328ebc6f2ba9802c1";
     char hex[sizeof(data) * 2 + 1];
